+++ http/src/main/scala/com/twitter/finatra/http/modules/ExceptionManagerModule.scala
@@ -2,7 +2,9 @@ package com.twitter.finatra.http.modules
 
 import com.google.inject.Provides
 import com.twitter.finagle.stats.StatsReceiver
-import com.twitter.finatra.http.exceptions.{ExceptionManager, DefaultExceptionMapper}
+import com.twitter.finatra.http.exceptions.ExceptionManager
+import com.twitter.finatra.http.internal.exceptions.ThrowableExceptionMapper
+import com.twitter.finatra.http.internal.exceptions.json.{CaseClassExceptionMapper, JsonParseExceptionMapper}
 import com.twitter.inject.{Injector, TwitterModule}
 import javax.inject.Singleton
 
@@ -12,9 +14,16 @@ object ExceptionManagerModule extends TwitterModule {
   @Singleton
   def providesExceptionManager(
     injector: Injector,
-    defaultExceptionMapper: DefaultExceptionMapper,
     statsReceiver: StatsReceiver
   ): ExceptionManager = {
-    new ExceptionManager(injector, defaultExceptionMapper, statsReceiver)
+    new ExceptionManager(injector, statsReceiver)
+  }
+
+  /** Add default Framework Exception Mappers */
+  override def singletonStartup(injector: Injector) {
+    val manager = injector.instance[ExceptionManager]
+    manager.add[ThrowableExceptionMapper]
+    manager.add[JsonParseExceptionMapper]
+    manager.add[CaseClassExceptionMapper]
   }
 }