+++ inject/inject-app/src/main/scala/com/twitter/inject/app/internal/TwitterTypeConvertersModule.scala
@@ -7,7 +7,7 @@ import org.joda.time.Duration
 private[app] object TwitterTypeConvertersModule extends TwitterModule {
 
   override def configure() {
-    addTypeConvertor[Duration](JodatimeDurationTypeConvertor)
-    addTypeConvertor[TwitterDuration](TwitterDurationTypeConvertor)
+    addTypeConverter[Duration](JodatimeDurationTypeConverter)
+    addTypeConverter[TwitterDuration](TwitterDurationTypeConverter)
   }
 }